# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hotel Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: flights
    description: Flight search and booking
  - name: hotels
    description: Hotel search and booking
  - name: bookings
    description: Booking management
  - name: offers
    description: Promotional offers
  - name: destinations
    description: Popular destinations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /flights/search:
    get:
      operationId: searchFlights
      tags: [flights]
      summary: Search for flights
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
        - name: passengers
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: class
          in: query
          required: false
          schema:
            type: string
            enum: [economy, business, first]
            default: economy
      responses:
        "200":
          description: List of available flights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flight"

  /hotels/search:
    get:
      operationId: searchHotels
      tags: [hotels]
      summary: Search for hotels
      parameters:
        - name: city
          in: query
          required: true
          schema:
            type: string
        - name: checkIn
          in: query
          required: true
          schema:
            type: string
        - name: checkOut
          in: query
          required: true
          schema:
            type: string
        - name: guests
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: rooms
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: List of available hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hotel"

  /hotels/{hotelId}:
    get:
      operationId: getHotelDetail
      tags: [hotels]
      summary: Get hotel detail
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hotel detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelDetail"
        "404":
          description: Hotel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /hotels/{hotelId}/rooms:
    get:
      operationId: getHotelRooms
      tags: [hotels]
      summary: Get available room types for a hotel
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: string
        - name: checkIn
          in: query
          required: false
          schema:
            type: string
        - name: checkOut
          in: query
          required: false
          schema:
            type: string
        - name: guests
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: List of available room types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomType"

  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get all bookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new hotel booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/{bookingId}:
    get:
      operationId: getBookingDetail
      tags: [bookings]
      summary: Get booking detail
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /offers:
    get:
      operationId: getOffers
      tags: [offers]
      summary: Get promotional offers
      responses:
        "200":
          description: List of promotional offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"

  /destinations/popular:
    get:
      operationId: getPopularDestinations
      tags: [destinations]
      summary: Get popular travel destinations
      responses:
        "200":
          description: List of popular destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destination"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Flight:
      type: object
      properties:
        id:
          type: string
        airline:
          type: string
        airlineCode:
          type: string
        flightNumber:
          type: string
        from:
          type: string
        fromCode:
          type: string
        to:
          type: string
        toCode:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: string
        duration:
          type: string
        stops:
          type: integer
        price:
          type: number
        class:
          type: string
        seatsAvailable:
          type: integer
        rating:
          type: number
      required:
        - id
        - airline
        - airlineCode
        - flightNumber
        - from
        - fromCode
        - to
        - toCode
        - departureTime
        - arrivalTime
        - duration
        - stops
        - price
        - class
        - seatsAvailable
        - rating

    Hotel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        address:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        pricePerNight:
          type: number
        amenities:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        category:
          type: string
        discount:
          type: integer
      required:
        - id
        - name
        - city
        - address
        - rating
        - reviewCount
        - pricePerNight
        - amenities
        - imageUrl
        - category
        - discount

    HotelDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        address:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        pricePerNight:
          type: number
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        discount:
          type: integer
        description:
          type: string
        policies:
          type: object
          properties:
            checkIn:
              type: string
            checkOut:
              type: string
            cancellation:
              type: string
            pets:
              type: string
          required:
            - checkIn
            - checkOut
            - cancellation
            - pets
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        latitude:
          type: number
        longitude:
          type: number
      required:
        - id
        - name
        - city
        - address
        - rating
        - reviewCount
        - pricePerNight
        - amenities
        - images
        - category
        - discount
        - description
        - policies
        - reviews
        - latitude
        - longitude

    Review:
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        avatar:
          type: string
        rating:
          type: number
        date:
          type: string
        title:
          type: string
        comment:
          type: string
        helpful:
          type: integer
      required:
        - id
        - author
        - avatar
        - rating
        - date
        - title
        - comment
        - helpful

    RoomType:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        maxGuests:
          type: integer
        size:
          type: string
        bedType:
          type: string
        pricePerNight:
          type: number
        originalPrice:
          type: number
        images:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        available:
          type: integer
        refundable:
          type: boolean
        mealIncluded:
          type: string
      required:
        - id
        - name
        - description
        - maxGuests
        - size
        - bedType
        - pricePerNight
        - originalPrice
        - images
        - amenities
        - available
        - refundable
        - mealIncluded

    CreateBookingRequest:
      type: object
      properties:
        hotelId:
          type: string
        hotelName:
          type: string
        roomTypeId:
          type: string
        roomTypeName:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        guests:
          type: integer
        rooms:
          type: integer
        pricePerNight:
          type: number
        totalPrice:
          type: number
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        specialRequests:
          type: string
      required:
        - hotelId
        - hotelName
        - roomTypeId
        - roomTypeName
        - checkIn
        - checkOut
        - guests
        - rooms
        - pricePerNight
        - totalPrice
        - guestName
        - guestEmail
        - guestPhone

    Booking:
      type: object
      properties:
        id:
          type: string
        bookingRef:
          type: string
        hotelId:
          type: string
        hotelName:
          type: string
        hotelImage:
          type: string
        city:
          type: string
        roomTypeId:
          type: string
        roomTypeName:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        guests:
          type: integer
        rooms:
          type: integer
        pricePerNight:
          type: number
        totalPrice:
          type: number
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        specialRequests:
          type: string
        status:
          type: string
          enum: [confirmed, pending, cancelled, completed]
        createdAt:
          type: string
        nights:
          type: integer
      required:
        - id
        - bookingRef
        - hotelId
        - hotelName
        - hotelImage
        - city
        - roomTypeId
        - roomTypeName
        - checkIn
        - checkOut
        - guests
        - rooms
        - pricePerNight
        - totalPrice
        - guestName
        - guestEmail
        - guestPhone
        - status
        - createdAt
        - nights

    Offer:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        discount:
          type: string
        code:
          type: string
        validTill:
          type: string
        imageUrl:
          type: string
        category:
          type: string
      required:
        - id
        - title
        - description
        - discount
        - code
        - validTill
        - imageUrl
        - category

    Destination:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        country:
          type: string
        imageUrl:
          type: string
        startingPrice:
          type: number
        description:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - country
        - imageUrl
        - startingPrice
        - description
        - tags