# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  hotelId: string;
  hotelName: string;
  roomTypeId: string;
  roomTypeName: string;
  checkIn: string;
  checkOut: string;
  guests: number;
  rooms: number;
  pricePerNight: number;
  totalPrice: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  specialRequests?: string;
}